+++ server/server.js
@@ -6,7 +6,6 @@ const apiRouter = require('./apiRouter');
 app.use(express.json());
 
 app.use('/api', apiRouter, (req, res) => {
-  console.log('sending from server ', res.locals.data);
   res.status(200).send(res.locals.data);
 })
